Moved sum function from mockAPI/utils to libs/utils and fixed up imports
